rename old_style to legacy
